requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
networkx>=3.0
python-louvain>=0.16
ruptures>=1.1.7
scikit-bio>=0.5.7
pyarrow>=11.0.0
fastparquet>=2023.0.0
zarr>=2.13.0
loguru>=0.6.0
pyyaml>=6.0
openpyxl>=3.1.0
duckdb>=0.8.0
xlrd>=2.0.0